test: migrate global_rule e2e
